+++ .scripts/Update-ClientModel.ps1
@@ -2,7 +2,7 @@
 
 function Remove-MultipartFormDataBinaryContent {
     $root = Split-Path $PSScriptRoot -Parent
-    $filePath = Join-Path -Path $root -ChildPath ".dotnet\src\Generated\Internal\MultiPartFormDataBinaryContent.cs"
+    $filePath = Join-Path -Path $root -ChildPath ".dotnet/src/Generated/Internal/MultiPartFormDataBinaryContent.cs"
     $file = Get-ChildItem -Path $filePath
 
     Write-Output "Removing $($file.FullName)"
@@ -12,7 +12,7 @@ function Remove-MultipartFormDataBinaryContent {
 
 function Remove-ChatMessageContentSerialization {
     $root = Split-Path $PSScriptRoot -Parent
-    $filePath = Join-Path -Path $root -ChildPath ".dotnet\src\Generated\Models\ChatMessageContent.Serialization.cs"
+    $filePath = Join-Path -Path $root -ChildPath ".dotnet/src/Generated/Models/ChatMessageContent.Serialization.cs"
     $file = Get-ChildItem -Path $filePath
 
     Write-Output "Removing $($file.FullName)"